LayoutTests/platform/glib/imported/w3c/web-platform-tests/cors/origin-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fnot-web-platform.test.web-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=ftp%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3Aweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=web-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%3F due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%2F due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%20%2F due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%23 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%2523 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%3A80 due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%2C%20* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=HTTP%3A%2F%2FWEB-PLATFORM.TEST%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=HTTP%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=- due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=** due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%2C* due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%2C due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%0B* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%0C* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%0B due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%0C due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%27*%27 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=%22*%22 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%20* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%20null due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*http%3A%2F%2F* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%20http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*%2C%20http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=null%20http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fexample.net due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=null due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=null%20* due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin= due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%2Fcors%2Forigin.htm due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800%2Fcors%2F due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fnot-web-platform.test%3A8800%2Fcors%2F due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Ftest%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2F.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2F*.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=&origin2=* due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*&origin2= due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*&origin2=* due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=&origin2=http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=*&origin2=http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
CONSOLE MESSAGE: Access-Control-Allow-Origin cannot contain more than one origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors//resources/cors-makeheader.py?origin=http%3A%2F%2Fweb-platform.test%3A8800&origin2=http%3A%2F%2Fweb-platform.test%3A8800 due to access control checks.
Access-Control-Allow-Origin handling


PASS Allow origin: *
PASS Allow origin: _*__
PASS Allow origin: [tab]*
PASS Allow origin: http://web-platform.test:8800
PASS Allow origin: _http://web-platform.test:8800
PASS Allow origin: _http://web-platform.test:8800___[tab]_
PASS Allow origin: [tab]http://web-platform.test:8800
PASS Disallow origin: http://not-web-platform.test.web-platform.test:8800
PASS Disallow origin: //web-platform.test:8800
PASS Disallow origin: ://web-platform.test:8800
PASS Disallow origin: ftp://web-platform.test:8800
PASS Disallow origin: http:://web-platform.test:8800
PASS Disallow origin: http:/web-platform.test:8800
PASS Disallow origin: http:web-platform.test:8800
PASS Disallow origin: web-platform.test:8800
PASS Disallow origin: http://web-platform.test:8800?
PASS Disallow origin: http://web-platform.test:8800/
PASS Disallow origin: http://web-platform.test:8800 /
PASS Disallow origin: http://web-platform.test:8800#
PASS Disallow origin: http://web-platform.test:8800%23
PASS Disallow origin: http://web-platform.test:8800:80
PASS Disallow origin: http://web-platform.test:8800, *
FAIL Disallow origin: http://web-platform.test:8800\0 assert_throws_dom: send function "function() { client.send() }" did not throw
PASS Disallow origin: HTTP://WEB-PLATFORM.TEST:8800
PASS Disallow origin: HTTP://web-platform.test:8800
PASS Disallow origin: -
PASS Disallow origin: **
PASS Disallow origin: ,*
PASS Disallow origin: *,
FAIL Disallow origin: \0* assert_throws_dom: send function "function() { client.send() }" did not throw
PASS Disallow origin: *
PASS Disallow origin: *
FAIL Disallow origin: *\0 assert_throws_dom: send function "function() { client.send() }" did not throw
PASS Disallow origin: *
PASS Disallow origin: *
PASS Disallow origin: '*'
PASS Disallow origin: "*"
PASS Disallow origin: * *
PASS Disallow origin: * null
PASS Disallow origin: *http://*
PASS Disallow origin: *http://web-platform.test:8800
PASS Disallow origin: * http://web-platform.test:8800
PASS Disallow origin: *, http://web-platform.test:8800
FAIL Disallow origin: \0http://web-platform.test:8800 assert_throws_dom: send function "function() { client.send() }" did not throw
PASS Disallow origin: null http://web-platform.test:8800
PASS Disallow origin: http://example.net
PASS Disallow origin: null
PASS Disallow origin: null *
PASS Disallow origin:
PASS Disallow origin: http://web-platform.test:8800/cors/origin.htm
PASS Disallow origin: http://web-platform.test:8800/cors/
PASS Disallow origin: http://not-web-platform.test:8800/cors/
PASS Disallow origin: test:8800
PASS Disallow origin: .test:8800
PASS Disallow origin: *.test:8800
PASS Disallow origin: http://test:8800
PASS Disallow origin: http://.test:8800
PASS Disallow origin: http://*.test:8800
PASS Disallow multiple headers (, *)
PASS Disallow multiple headers (*, )
PASS Disallow multiple headers (*, *)
PASS Disallow multiple headers (, http://web-platform.test:8800)
PASS Disallow multiple headers (*, http://web-platform.test:8800)
PASS Disallow multiple headers (http://web-platform.test:8800, http://web-platform.test:8800)
